+++ kernel/src/pci/express.rs
@@ -9,9 +9,12 @@ use x86_64::{
 
 use crate::{
     boot::acpi::sdt::mcfg::ParsedConfigurationSpace, mem::page_alloc::pmap::PhysicalMapping,
+    pci::Class,
 };
 
-use super::commons::{Bar, Device};
+use super::commons::{
+    Bar, BdfAddress, ConfigAddressValue, Csp, Device, MemoryBarType, RegisterOffset,
+};
 
 mod msi;
 
@@ -19,6 +22,7 @@ static PCIE_HANDLER: MUMcsLock<PciExpressHandler> = MUMcsLock::uninit();
 
 pub struct PciExpressHandler {
     configuration_spaces: &'static [ParsedConfigurationSpace],
+    physical_mappings: Vec<PhysicalMapping<Size2MiB>>,
     devices: Vec<Device>,
 }
 
@@ -44,54 +48,154 @@ pub fn init() {
 impl PciExpressHandler {
     #[must_use]
     #[inline]
-    pub const fn new(configuration_spaces: &'static [ParsedConfigurationSpace]) -> Self {
+    pub fn new(configuration_spaces: &'static [ParsedConfigurationSpace]) -> Self {
+        let physical_mappings = configuration_spaces
+            .iter()
+            .map(|cs| {
+                let length = usize::try_from(
+                    cs.address_range().end().as_u64() - cs.address_range().start().as_u64(),
+                )
+                .unwrap();
+
+                let flags = PageTableFlags::PRESENT
+                    | PageTableFlags::WRITABLE
+                    | PageTableFlags::NO_EXECUTE
+                    | PageTableFlags::NO_CACHE;
+                PhysicalMapping::<Size2MiB>::new(*cs.address_range().start(), length, flags)
+            })
+            .collect::<Vec<_>>();
+
         Self {
             configuration_spaces,
+            physical_mappings,
             devices: Vec::new(),
         }
     }
 
     fn update_devices(&mut self) {
-        for cs in self.configuration_spaces {
-            let start_paddr = Self::build_paddr(cs.offset(), cs.start_pci_bus_number(), 0, 0);
-            let end_paddr = Self::build_paddr(cs.offset(), cs.end_pci_bus_number(), 31, 7);
+        self.devices.clear();
 
-            let length = usize::try_from(end_paddr.as_u64() - start_paddr.as_u64()).unwrap();
-
-            let flags = PageTableFlags::PRESENT
-                | PageTableFlags::WRITABLE
-                | PageTableFlags::NO_EXECUTE
-                | PageTableFlags::NO_CACHE;
-            let pmap = PhysicalMapping::<Size2MiB>::new(start_paddr, length, flags);
-
+        // Brute-force scan
+        for (cs, pmap) in self
+            .configuration_spaces
+            .iter()
+            .zip(&self.physical_mappings)
+        {
             for bus in cs.start_pci_bus_number()..=cs.end_pci_bus_number() {
                 for dev in 0..=31 {
-                    self.devices.push(Self::scan_device(&pmap, cs.offset(), bus, dev));
+                    if let Some(device) = Self::scan_device(
+                        pmap,
+                        cs,
+                        ConfigAddressValue::new(bus, dev, 0, RegisterOffset::VendorId as u8),
+                    ) {
+                        self.devices.push(device);
+                    }
                 }
             }
         }
     }
 
-    fn scan_device(pmap: &PhysicalMapping<Size2MiB>, offset: u64, bus: u8, dev: u8) -> super::commons::Device {
-        // TODO: Scan is performed the same way as for PCI devices
-        // except it is done using MMIO and not I/O ports.
-        let paddr = Self::build_paddr(offset, bus, dev, 0);
+    fn scan_device(
+        pmap: &PhysicalMapping<Size2MiB>,
+        cs: &ParsedConfigurationSpace,
+        address: ConfigAddressValue,
+    ) -> Option<Device> {
+        let (device, vendor) = {
+            let reg = ConfigAddressValue {
+                register_offset: RegisterOffset::VendorId as u8,
+                ..address
+            };
+            let paddr = Self::build_paddr(cs.offset(), reg);
+            let vaddr = pmap.translate(paddr)?;
+            let value = unsafe { vaddr.as_ptr::<u32>().read() };
 
-        let vaddr = pmap.translate(paddr).unwrap();
-        let vendor = unsafe { vaddr.as_ptr::<u16>().read() };
-        if vendor != u16::MAX {
-            crate::debug!("PCIe device found with vendor {}", vendor);
-        }
+            if value & 0xFFFF == u32::from(u16::MAX) {
+                return None;
+            }
 
-        todo!("PCIe device scanning");
+            (
+                u16::try_from(value >> 16).unwrap(),
+                u16::try_from(value & 0xFFFF).unwrap(),
+            )
+        };
+
+        let (class, subclass, prog_if, revision) = {
+            let reg = ConfigAddressValue {
+                register_offset: RegisterOffset::RevisionId as u8,
+                ..address
+            };
+            let paddr = Self::build_paddr(cs.offset(), reg);
+            let vaddr = pmap.translate(paddr)?;
+            let value = unsafe { vaddr.as_ptr::<u32>().read() };
+            (
+                Class::from(u8::try_from(value >> 24).unwrap()),
+                u8::try_from((value >> 16) & 0xFF).unwrap(),
+                u8::try_from((value >> 8) & 0xFF).unwrap(),
+                u8::try_from(value & 0xFF).ok().unwrap(),
+            )
+        };
+
+        let functions = Self::find_function_count(pmap, cs.offset(), address);
+
+        Some(Device {
+            id: device,
+            vendor_id: vendor,
+            bdf: address.bdf,
+            functions,
+            csp: Csp::new(class, subclass, prog_if),
+            revision,
+            segment_group_number: cs.segment_group_number(),
+        })
     }
 
-    fn build_paddr(offset: u64, bus: u8, dev: u8, func: u8) -> PhysAddr {
-        let bus = u64::from(bus);
-        let dev = u64::from(dev);
-        let func = u64::from(func);
+    fn find_function_count(
+        pmap: &PhysicalMapping<Size2MiB>,
+        offset: u64,
+        address: ConfigAddressValue,
+    ) -> u8 {
+        let multifonction = {
+            let reg = ConfigAddressValue {
+                register_offset: RegisterOffset::HeaderType as u8,
+                ..address
+            };
+            let paddr = Self::build_paddr(offset, reg);
+            let vaddr = pmap.translate(paddr).unwrap();
+            let value = unsafe { vaddr.as_ptr::<u8>().read() };
 
-        let paddr = offset + (bus << 20) + (dev << 15) + (func << 12);
+            value >> 7 == 1
+        };
+
+        if !multifonction {
+            return 1;
+        }
+
+        u8::try_from(
+            (1..8)
+                .filter(|&func| {
+                    let reg = ConfigAddressValue {
+                        bdf: BdfAddress::new(address.bdf.bus(), address.bdf.device(), func),
+                        register_offset: RegisterOffset::VendorId as u8,
+                        ..address
+                    };
+                    let paddr = Self::build_paddr(offset, reg);
+                    let vaddr = pmap.translate(paddr).unwrap();
+
+                    // Vendor ID is 0xFFFF if function is unsupported
+                    unsafe { vaddr.as_ptr::<u16>().read() != u16::MAX }
+                })
+                .count(),
+        )
+        .unwrap()
+            + 1
+    }
+
+    fn build_paddr(offset: u64, address: ConfigAddressValue) -> PhysAddr {
+        let bus = u64::from(address.bdf.bus());
+        let dev = u64::from(address.bdf.device());
+        let func = u64::from(address.bdf.function());
+        let reg = u64::from(address.register_offset);
+
+        let paddr = offset + (bus << 20) + (dev << 15) + (func << 12) + reg;
         PhysAddr::new(paddr)
     }
 }
@@ -104,9 +208,79 @@ impl super::PciHandler for PciExpressHandler {
     fn read_bar(
         &mut self,
         device: &super::commons::Device,
-        bar: u8,
+        bar_number: u8,
     ) -> Option<super::commons::Bar> {
-        None
+        let bar_reg_offset = match bar_number {
+            0 => RegisterOffset::Bar0,
+            1 => RegisterOffset::Bar1,
+            2 => RegisterOffset::Bar2,
+            3 => RegisterOffset::Bar3,
+            4 => RegisterOffset::Bar4,
+            5 => RegisterOffset::Bar5,
+            _ => return None,
+        } as u8;
+        let reg = ConfigAddressValue::new(
+            device.bdf().bus(),
+            device.bdf().device(),
+            device.bdf().function(),
+            bar_reg_offset,
+        );
+
+        let (cs, pmap) = {
+            let cs_index = self
+                .configuration_spaces
+                .iter()
+                .position(|cs| cs.segment_group_number() == device.segment_group_number)
+                .unwrap();
+
+            (
+                &self.configuration_spaces[cs_index],
+                &self.physical_mappings[cs_index],
+            )
+        };
+
+        let paddr = Self::build_paddr(cs.offset(), reg);
+        let vaddr = pmap.translate(paddr).unwrap();
+
+        let bar = unsafe { vaddr.as_ptr::<u32>().read() };
+        if bar == u32::MAX {
+            return None;
+        }
+
+        let upper_value = if bar & 1 == 0 // Memory BAR
+            && MemoryBarType::try_from((bar >> 1) & 0b11).unwrap() == MemoryBarType::Qword
+        {
+            let bar_reg_offset = match bar_number + 1 {
+                0 => RegisterOffset::Bar0,
+                1 => RegisterOffset::Bar1,
+                2 => RegisterOffset::Bar2,
+                3 => RegisterOffset::Bar3,
+                4 => RegisterOffset::Bar4,
+                5 => RegisterOffset::Bar5,
+                _ => panic!("PCI: Invalid BAR number"),
+            } as u8;
+            let bar_reg = ConfigAddressValue::new(
+                device.bdf().bus(),
+                device.bdf().device(),
+                device.bdf().function(),
+                bar_reg_offset,
+            );
+
+            let paddr = Self::build_paddr(cs.offset(), bar_reg);
+            let vaddr = pmap.translate(paddr).unwrap();
+
+            let bar = unsafe { vaddr.as_ptr::<u32>().read() };
+            if bar == u32::MAX {
+                return None;
+            }
+            bar
+        } else {
+            0
+        };
+
+        Some(Bar::from_raw(
+            u64::from(bar) | (u64::from(upper_value) << 32),
+        ))
     }
 }
